Changed as many OpenGl calls as possible to use the GlStateManager class
The GlStateManager class seems to be used to track the state of OpenGL (something raw calls cannot do), so it is presumably best practice to use this class. This presumption is reinforced by the fact that the text rendering problems went away after switching from GL11 calls to GlStateManager calls.