Align tooltip drawing method with Minecrafts method
The method used by vanilla can be found in the
net.minecraftforge.fml.client.config.GuiUtils class and it very
closely resembles the one used in the AntiqueAtlas GuiComponent
class.

I've cleaned up this version to use the same deobfuscated names
as the Minecraft one and added the firing of the Events currently
being fired by forge.

This allows mods like Colored Tooltips to modify the background
color of the Antique Atlas tooltips.

Ping @Darkosto