Watchers should ignore irrelevant MapGenStructureData tags (AntiqueAtlasTeam/AntiqueAtlas#168) (#169)
Well, even if I can't recreate the bug, I see nothing wrong with an extra check on the coordinates. It might save some time debugging in the future. I'm pretty sure the checks failed because the *build.gradle* file was messed up at one point, and I forgot to commit the fixes, so this should be save to merge.