Only run Atlas updates on the server.
This used to be the case until way back in 4.4.5, which, according to the commit message, fixed an ominous bug. I can't track down the exact bug, so I'm just hoping that this is fine now.

At the very least, this hides AA being a resource hog from clients O:)

Co-Authored-By: tyra314 <16997177+tyra314@users.noreply.github.com>